TopologyDisplayView: The tab type remains the same (if possible) when the user selects a new node.
The v-btn-toggle was connecting its v-model to the v-btn keys. This caused the model to become meaningless when the node was changed. Using the tab.type as the button value allows the toggle group model to behave as expected and simplifies the code by removing the need for seperate activeTab and activeTabType variables, with activeTab now containing the tab.type string.

Some additional cleaning up was performed